// types/mb-ribbon.ts
export interface MbRibbon {
  id: number;
  tracking: string;
  user_id: number;
  created_at: string;
  updated_at: string;
  complained: boolean;
  order?: {
    id: number;
    order_id: string;
    status: string;
    channel: string;
    store: string;
    buyer: string;
    courier: string;
    tracking: string;
    complained: boolean;
    picked_by: string;
    picked_at: string;
    created_at: string;
    updated_at: string;
    order_details: unknown[];
  };
  user?: {
    id: number;
    username: string;
    email: string;
    full_name: string;
    is_active: boolean;
    created_at: string;
    updated_at: string;
    roles: Array<{
      id: number;
      name: string;
      description: string;
      assigned_by: string;
      assigned_at: string;
    }>;
  };
}